+++ src/cli/node/start/start.js
@@ -95,6 +95,12 @@ async function start(chainName, dryRun = false, allowMultiChain) {
       throw new Error(`${errMessage.status}: ${errMessage.message}`);
     }
 
+    // check forge process again
+    const tmpResult = await isForgeStarted(chainName);
+    if (!tmpResult) {
+      throw new Error('Start forge failed');
+    }
+
     spinner.succeed(`Chain ${chalk.yellow(chainName)} successfully started`);
     if (config.get('forge.web.enabled')) {
       await startWeb();