+++ src/core/libs/common.js
@@ -170,6 +170,9 @@ const makeNativeCommand = ({ binPath, subcommand = 'daemon', env = '' }) => {
   return command;
 };
 
+const getChainGraphQLHost = config =>
+  process.env.FORGE_GQL_ENDPOINT || `http://127.0.0.1:${config.get('forge.web.port', 8210)}/api`;
+
 module.exports = {
   DEFAULT_CHAIN_NAME_RETURN,
   applyForgeVersion,
@@ -180,6 +183,7 @@ module.exports = {
   },
   fetchPackageJSON,
   fetchLatestCLIVersion,
+  getChainGraphQLHost,
   getChainVersion,
   getDefaultChainNameHandlerByChains,
   getMinSupportForgeVersion,