+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.40.3 (November 04, 2019)
+
+- chore: check if forge process exists before start web when run `forge start`
+- improvement: validate ugprade tx during upgrade
+
 ## 0.40.2 (November 02, 2019)
 
 - fix: create chain failed